+++ src/async.js
@@ -1,3 +1,10 @@
+/*-------------------------------------------------------------------------------------------
+ * validatorjs
+ *
+ * Copyright (c) 2021 Mike Erickson / Codedungeon.  All rights reserved.
+ * Licensed under the MIT license.  See LICENSE in the project root for license information.
+ * -----------------------------------------------------------------------------------------*/
+
 "use strict";
 
 function AsyncResolvers(onFailedOne, onResolvedAll) {